LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/exclude-sticky-expected.txt

sticky
content

FAIL Sticky-positioned headers shouldn't be chosen as scroll anchors (we should use 'content' instead) assert_equals: expected 200 but got 150
